Unify common test behaviour
- Prevents issue with RunEngine event_loop not being available
- Reduces duplication
- Expandable to new beamlines
